feat: moved get_utc_time over to Socket, change time periodically
