+++ crates/comm/src/socket/mod.rs
@@ -10,6 +10,10 @@ use std::{
     time::Duration,
 };
 
+use prost_types::Timestamp;
+use rsntp::AsyncSntpClient;
+use chrono::Datelike;
+
 use rand::{rngs::OsRng, seq::IteratorRandom};
 use string_protocol::{MessageType, ProtocolPacket};
 use tokio::{
@@ -77,6 +81,8 @@ pub struct Socket {
     pub username: String,
     /// Channel used to send gossip
     pub gossip_tx: mpsc::Sender<Gossip>,
+	/// current timestamp used by [Peer] to get available peers
+	pub curr_time: Arc<RwLock<Timestamp>>,
 }
 
 impl Socket {
@@ -95,6 +101,8 @@ impl Socket {
         let crypto = Arc::new(RwLock::new(Crypto::new(secret_key.clone())));
 
         let (gossip_tx, gossip_rx) = mpsc::channel(CHANNEL_SIZE);
+		
+		let curr_time = Arc::new(RwLock::new(Self::get_utc_time().await?));
 
         // start the outbound worker
         span!(tracing::Level::INFO, "socket::outbound")
@@ -106,7 +114,7 @@ impl Socket {
 
         // start the perodic worker
         span!(tracing::Level::INFO, "socket::periodic")
-            .in_scope(|| start_periodic_worker(peers.clone()));
+            .in_scope(|| start_periodic_worker(peers.clone(), curr_time.clone()));
 
         if secret_key.details.users.len() != 1 {
             // Why do we have a weird number of users
@@ -121,6 +129,7 @@ impl Socket {
             crypto,
             username,
             gossip_tx,
+			curr_time,
         })
     }
 
@@ -139,6 +148,7 @@ impl Socket {
             self.username.clone(),
             self.gossip_tx.clone(),
             fingerprint,
+			self.curr_time.clone(),
             initiate,
         ).await?;
 
@@ -284,6 +294,18 @@ impl Socket {
         }
     }
 
+    async fn get_utc_time() -> Result<Timestamp, SocketError> {
+        let client = AsyncSntpClient::new();
+        let result = client.synchronize("pool.ntp.org").await?;
+
+        let curr_time = result.datetime().into_chrono_datetime()?;
+        // unwrap here is safe (month is <= 12 and day is <= 31)
+        Ok(Timestamp::date(
+            curr_time.year().into(),
+            curr_time.month().try_into().unwrap(),
+            curr_time.day().try_into().unwrap(),
+        )?)
+    }
 }
 
 /// Start the outbound network worker.
@@ -430,7 +452,7 @@ fn start_gossip_worker(
 }
 
 /// Starts a background worker than can do certain chores at regular intervals
-fn start_periodic_worker(peers: Arc<RwLock<HashMap<SocketAddr, Peer>>>) {
+fn start_periodic_worker(peers: Arc<RwLock<HashMap<SocketAddr, Peer>>>, curr_time: Arc<RwLock<Timestamp>>) {
     tokio::spawn(async move {
         loop {
             tokio::time::sleep(Duration::from_millis(5000)).await;
@@ -438,6 +460,15 @@ fn start_periodic_worker(peers: Arc<RwLock<HashMap<SocketAddr, Peer>>>) {
 			for peer in peers_write.values_mut(){
 				try_continue!(peer.request_available_peers().await);
 			}
+
+			// periodically, update time
+			let updated_time = Socket::get_utc_time().await;
+			// skip iteration if there's an error
+			// otherwise, get the value inside. 
+			match updated_time{
+				Ok(updated_time) => *curr_time.write().await = updated_time,
+				_ => continue
+			};
         }
     });
 }
